+++ priv/repo/migrations/20211129012408_add_df_content_to_execution_groups.exs
@@ -1,0 +1,9 @@
+defmodule TdDd.Repo.Migrations.AddDfContentToExecutionGroups do
+  use Ecto.Migration
+
+  def change do
+    alter table("execution_groups") do
+      add :df_content, :map
+    end
+  end
+end